chore: move plugins into the `apps` folder
