+++ src/apps/halofi/plugin.e2e.ts
