+++ src/apps/locked-celo/abis/locked-gold.ts
