+++ src/apps/locked-celo/plugin.e2e.ts
